+++ iridauploader/model/sequencing_run.py
@@ -1,3 +1,12 @@
+"""
+A SequencingRun contains all required information from a run
+
+project_list: list of Project objects
+metadata: dict of key value pairs for additional data
+    the metadata dict requires the key `layoutType` with one of ['PAIRED_END', 'SINGLE_END']
+assemblies: boolean, True if uploading assemblies
+sequencing_run_type: string for path uploading to.
+"""
 from cerberus import Validator, TypeDefinition
 
 from iridauploader.model.project import Project
@@ -28,13 +37,18 @@ class SequencingRun:
         '_assemblies': {
             'type': 'boolean',
             'required': True
+        },
+        '_sequencing_run_type': {
+            'type': 'string',
+            'required': True
         }
     }
 
-    def __init__(self, metadata, project_list, assemblies=False):
+    def __init__(self, metadata, project_list, sequencing_run_type, assemblies=False):
         self._project_list = project_list
         self._metadata = metadata
         self._assemblies = assemblies
+        self._sequencing_run_type = sequencing_run_type
 
     @property
     def metadata(self):
@@ -51,6 +65,14 @@ class SequencingRun:
     @project_list.setter
     def project_list(self, p_list):
         self._project_list = p_list
+
+    @property
+    def sequencing_run_type(self):
+        return self._sequencing_run_type
+
+    @sequencing_run_type.setter
+    def upload_route_string(self, sequencing_run_type):
+        self._sequencing_run_type = sequencing_run_type
 
     @property
     def assemblies(self):