+++ iridauploader/parsers/__init__.py
@@ -1,4 +1,5 @@
-from iridauploader.parsers.parsers import Parser
+from iridauploader.parsers.base_parser import BaseParser
 from iridauploader.parsers.parsers import supported_parsers
+from iridauploader.parsers.parsers import parser_factory
 from iridauploader.parsers import exceptions
 from iridauploader.parsers import common