+++ iridauploader/tests/parsers/directory/test_parser.py
@@ -29,7 +29,7 @@ class TestFindRuns(unittest.TestCase):
         dir_3 = path.join(directory, "third")
         correct_dirs = [dir_1, dir_2, dir_3]
 
-        res = Parser.find_runs(directory)
+        res = Parser().find_runs(directory)
 
         self.assertIn(res[0].directory, correct_dirs)
         self.assertIn(res[1].directory, correct_dirs)
@@ -50,7 +50,7 @@ class TestFindRuns(unittest.TestCase):
         """
         directory = path.join(path_to_module, "no_dirs")
 
-        res = Parser.find_runs(directory)
+        res = Parser().find_runs(directory)
 
         self.assertEqual(res, [])
 
@@ -70,7 +70,7 @@ class TestFindSingleRun(unittest.TestCase):
         """
         directory = path.join(path_to_module, "three_dirs")
 
-        res = Parser.find_single_run(directory)
+        res = Parser().find_single_run(directory)
 
         self.assertEqual(type(res), model.DirectoryStatus)
         self.assertEqual(res.status, "invalid")
@@ -83,7 +83,7 @@ class TestFindSingleRun(unittest.TestCase):
         """
         directory = path.join(path_to_module, "no_dirs")
 
-        res = Parser.find_single_run(directory)
+        res = Parser().find_single_run(directory)
 
         self.assertEqual(type(res), model.DirectoryStatus)
         self.assertEqual(res.status, "invalid")
@@ -96,7 +96,7 @@ class TestFindSingleRun(unittest.TestCase):
         """
         directory = path.join(path_to_module, "three_dirs", "first")
 
-        res = Parser.find_single_run(directory)
+        res = Parser().find_single_run(directory)
 
         self.assertEqual(type(res), model.DirectoryStatus)
         self.assertEqual(res.status, "new")
@@ -119,7 +119,7 @@ class TestGetSampleSheet(unittest.TestCase):
         directory = path.join(path_to_module, "three_dirs", "first")
         file_path = path.join(directory, "SampleList.csv")
 
-        res = Parser.get_sample_sheet(directory)
+        res = Parser().get_sample_sheet(directory)
 
         self.assertEqual(res, file_path)
 
@@ -131,7 +131,7 @@ class TestGetSampleSheet(unittest.TestCase):
         directory = path.join(path_to_module, "three_dirs", "third")
 
         with self.assertRaises(DirectoryError) as context:
-            Parser.get_sample_sheet(directory)
+            Parser().get_sample_sheet(directory)
 
         self.assertEqual(context.exception.directory, directory)
 
@@ -143,7 +143,7 @@ class TestGetSampleSheet(unittest.TestCase):
         directory = path.join(path_to_module, "inaccessible_dir")
 
         with self.assertRaises(DirectoryError) as context:
-            Parser.get_sample_sheet(directory)
+            Parser().get_sample_sheet(directory)
 
         self.assertEqual(context.exception.directory, directory)
 
@@ -164,7 +164,7 @@ class TestGetSequencingRun(unittest.TestCase):
         sample_sheet = path.join(path_to_module, "invalid_sample_sheet", "SampleList.csv")
 
         with self.assertRaises(ValidationError) as context:
-            Parser.get_sequencing_run(sample_sheet)
+            Parser().get_sequencing_run(sample_sheet)
 
         validation_result = context.exception.validation_result
         self.assertEqual(type(validation_result), model.ValidationResult)
@@ -179,7 +179,7 @@ class TestGetSequencingRun(unittest.TestCase):
         """
         sample_sheet = path.join(path_to_module, "fake_dir_data", "SampleList.csv")
 
-        res = Parser.get_sequencing_run(sample_sheet)
+        res = Parser().get_sequencing_run(sample_sheet)
 
         self.assertEqual(type(res), model.SequencingRun)
 
@@ -191,7 +191,7 @@ class TestGetSequencingRun(unittest.TestCase):
         sheet_file = path.join(path_to_module, "fake_dir_data",
                                "SampleList.csv")
 
-        sequencing_run = Parser.get_sequencing_run(sheet_file)
+        sequencing_run = Parser().get_sequencing_run(sheet_file)
 
         # Returns a SequencingRun
         self.assertEqual(type(sequencing_run), model.SequencingRun)
@@ -216,7 +216,7 @@ class TestGetSequencingRun(unittest.TestCase):
         sheet_file = path.join(path_to_module, "fake_dir_data",
                                "SampleList_with_space.csv")
 
-        sequencing_run = Parser.get_sequencing_run(sheet_file)
+        sequencing_run = Parser().get_sequencing_run(sheet_file)
 
         # Returns a SequencingRun
         self.assertEqual(type(sequencing_run), model.SequencingRun)
@@ -256,7 +256,7 @@ class TestGetSequencingRun(unittest.TestCase):
             "",
         )
 
-        res = Parser.get_sequencing_run(sheet_file)
+        res = Parser().get_sequencing_run(sheet_file)
 
         self.assertEqual(res.metadata, {'layoutType': 'PAIRED_END'})
         self.assertEqual(res.project_list[0].id, "75")